reorganizing into client and b-store
