+++ client/lib/storage/general-settings-db-store.ts
@@ -1,7 +1,6 @@
 import type { SqliteAdapter } from './sqlite-adapter'
 import {
   DEFAULT_GENERAL_SETTINGS,
-  GENERAL_SETTINGS_CHANGE_EVENT,
   type FontSetting,
   type GeneralSettingsState,
   type SearchScopeSetting,
@@ -15,26 +14,12 @@ type GeneralSettingsRow = {
   setting_value: string
 }
 
-type GeneralSettingsApi = {
-  getGeneralSettings: () => Promise<GeneralSettingsState>
-  setGeneralSettings: (settings: GeneralSettingsState) => Promise<GeneralSettingsState>
-  updateGeneralSetting: (
-    key: PersistedGeneralSettingKey,
-    value: GeneralSettingsState[PersistedGeneralSettingKey]
-  ) => Promise<GeneralSettingsState>
-  resetGeneralSettings: () => Promise<GeneralSettingsState>
-}
-
 const schemaSql = `
 CREATE TABLE IF NOT EXISTS general_settings (
   setting_key TEXT PRIMARY KEY,
   setting_value TEXT NOT NULL
 );
 `
-
-const getGeneralSettingsApi = (): GeneralSettingsApi => {
-  return (window.conveyor as any).generalSettings as GeneralSettingsApi
-}
 
 const parseBoolean = (value: string, fallback: boolean): boolean => {
   if (value === 'true' || value === '1') return true
@@ -58,24 +43,6 @@ const valueToStorage = (value: GeneralSettingsState[PersistedGeneralSettingKey])
   return typeof value === 'boolean' ? String(value) : value
 }
 
-const coerceValue = <K extends PersistedGeneralSettingKey>(
-  key: K,
-  rawValue: string,
-  fallback: GeneralSettingsState[K]
-): GeneralSettingsState[K] => {
-  if (key === 'launch-on-startup') {
-    return parseBoolean(rawValue, fallback as boolean) as GeneralSettingsState[K]
-  }
-  if (key === 'theme') {
-    return parseTheme(rawValue, fallback as ThemeSetting) as GeneralSettingsState[K]
-  }
-  if (key === 'font') {
-    return parseFont(rawValue, fallback as FontSetting) as GeneralSettingsState[K]
-  }
-
-  return parseScope(rawValue, fallback as SearchScopeSetting) as GeneralSettingsState[K]
-}
-
 const runUpsert = (
   adapter: SqliteAdapter,
   key: PersistedGeneralSettingKey,
@@ -90,36 +57,6 @@ const runUpsert = (
   )
 }
 
-export const loadGeneralSettings = async (): Promise<GeneralSettingsState> => {
-  try {
-    return await getGeneralSettingsApi().getGeneralSettings()
-  } catch (error) {
-    console.error('Failed to load general settings from DB:', error)
-    return { ...DEFAULT_GENERAL_SETTINGS }
-  }
-}
-
-export const saveGeneralSettings = async (settings: GeneralSettingsState): Promise<GeneralSettingsState> => {
-  const result = await getGeneralSettingsApi().setGeneralSettings(settings)
-  window.dispatchEvent(new CustomEvent(GENERAL_SETTINGS_CHANGE_EVENT))
-  return result
-}
-
-export const updateGeneralSetting = async <K extends PersistedGeneralSettingKey>(
-  key: K,
-  value: GeneralSettingsState[K]
-): Promise<GeneralSettingsState> => {
-  const result = await getGeneralSettingsApi().updateGeneralSetting(key, value)
-  window.dispatchEvent(new CustomEvent(GENERAL_SETTINGS_CHANGE_EVENT))
-  return result
-}
-
-export const resetGeneralSettings = async (): Promise<GeneralSettingsState> => {
-  const result = await getGeneralSettingsApi().resetGeneralSettings()
-  window.dispatchEvent(new CustomEvent(GENERAL_SETTINGS_CHANGE_EVENT))
-  return result
-}
-
 export const createGeneralSettingsStore = (adapter: SqliteAdapter) => {
   const init = () => {
     adapter.exec(schemaSql)
@@ -130,12 +67,23 @@ export const createGeneralSettingsStore = (adapter: SqliteAdapter) => {
     const settings: GeneralSettingsState = { ...DEFAULT_GENERAL_SETTINGS }
 
     for (const row of rows) {
-      if (!(row.setting_key in settings)) {
-        continue
+      switch (row.setting_key) {
+        case 'launch-on-startup':
+          settings['launch-on-startup'] = parseBoolean(
+            row.setting_value,
+            DEFAULT_GENERAL_SETTINGS['launch-on-startup']
+          )
+          break
+        case 'theme':
+          settings.theme = parseTheme(row.setting_value, DEFAULT_GENERAL_SETTINGS.theme)
+          break
+        case 'font':
+          settings.font = parseFont(row.setting_value, DEFAULT_GENERAL_SETTINGS.font)
+          break
+        case 'scope':
+          settings.scope = parseScope(row.setting_value, DEFAULT_GENERAL_SETTINGS.scope)
+          break
       }
-
-      const key = row.setting_key as PersistedGeneralSettingKey
-      settings[key] = coerceValue(key, row.setting_value, settings[key])
     }
 
     return settings
@@ -154,7 +102,7 @@ export const createGeneralSettingsStore = (adapter: SqliteAdapter) => {
     return getGeneralSettings()
   }
 
-  const updateGeneralSettingValue = <K extends PersistedGeneralSettingKey>(
+  const updateGeneralSetting = <K extends PersistedGeneralSettingKey>(
     key: K,
     value: GeneralSettingsState[K]
   ): GeneralSettingsState => {
@@ -162,7 +110,7 @@ export const createGeneralSettingsStore = (adapter: SqliteAdapter) => {
     return getGeneralSettings()
   }
 
-  const resetGeneralSettingsValues = (): GeneralSettingsState => {
+  const resetGeneralSettings = (): GeneralSettingsState => {
     adapter.exec('DELETE FROM general_settings')
     return { ...DEFAULT_GENERAL_SETTINGS }
   }
@@ -171,8 +119,8 @@ export const createGeneralSettingsStore = (adapter: SqliteAdapter) => {
     init,
     getGeneralSettings,
     setGeneralSettings,
-    updateGeneralSetting: updateGeneralSettingValue,
-    resetGeneralSettings: resetGeneralSettingsValues,
+    updateGeneralSetting,
+    resetGeneralSettings,
     close: () => adapter.close(),
   }
 }
